redo how capture headers are handled
Create a pkthdr class that wraps struct pcap_pkthdr and use this class
whereever packet headers are used.  Even though the C structure has a
microsecond field, the new class defines an nsec property
(nanoseconds) instead of usec (microseconds).  This is in anticipation
of taking advantage of libpcap's newly added nanosecond precision
support in a future commit.  (libpcap recently extended its API to
turn on an option to abuse the microsecond field to store nanoseconds
instead of microseconds.)