delete pcap.readpkts() method
The libpcap documentation says that the packet data is not guaranteed
to be valid after the pcap_dispatch() method returns, and readpkts()
doesn't make a copy of the data, so simply remove the method.

This could be fixed by making a copy of the packet data in the
__add_pkts() method, but the performance implications and risk of
unbounded memory consumption (when reading a pcap file) make it easier
to push the burden of collecting packets in memory onto the user.